requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_ipyw_jlab_nb_ext_conf=0.1.0=py37_0
_libgcc_mutex=0.1=main
alabaster=0.7.12=py37_0
amply=0.1.4=pypi_0
anaconda=2018.12=py37_0
anaconda-client=1.7.2=py37_0
anaconda-navigator=1.9.6=py37_0
anaconda-project=0.8.2=py37_0
asn1crypto=0.24.0=py37_0
astroid=2.1.0=py37_0
astropy=3.1=py37h7b6447c_0
atomicwrites=1.2.1=py37_0
attrs=18.2.0=py37h28b3542_0
babel=2.6.0=py37_0
backcall=0.1.0=py37_0
backports=1.0=py37_1
backports.os=0.1.1=py37_0
backports.shutil_get_terminal_size=1.0.0=py37_2
beautifulsoup4=4.6.3=py37_0
bitarray=0.8.3=py37h14c3975_0
bkcharts=0.2=py37_0
blas=1.0=mkl
blaze=0.11.3=py37_0
bleach=3.0.2=py37_0
blosc=1.14.4=hdbcaa40_0
bokeh=1.0.2=py37_0
boto=2.49.0=py37_0
bottleneck=1.2.1=py37h035aef0_1
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.25=h06a4308_1
cairo=1.14.12=h8948797_3
certifi=2021.5.30=py37h06a4308_0
cffi=1.11.5=py37he75722e_1
chardet=3.0.4=py37_1
click=7.0=py37_0
cloudpickle=0.6.1=py37_0
clyent=1.2.2=py37_1
colorama=0.4.1=py37_0
conda=4.10.1=py37h06a4308_1
conda-build=3.17.6=py37_0
conda-env=2.6.0=1
conda-package-handling=1.7.3=py37h27cfd23_1
conda-verify=3.1.1=py37_0
contextlib2=0.5.5=py37_0
cryptography=2.4.2=py37h1ba5d50_0
cudatoolkit=11.1.74=h6bb024c_0
curl=7.63.0=hbc83047_1000
cycler=0.10.0=py37_0
cython=0.29.2=py37he6710b0_0
cytoolz=0.9.0.1=py37h14c3975_1
dask=1.0.0=py37_0
dask-core=1.0.0=py37_0
datashape=0.5.4=py37_1
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py37_0
defusedxml=0.5.0=py37_1
distributed=1.25.1=py37_0
docutils=0.14=py37_0
entrypoints=0.2.3=py37_2
et_xmlfile=1.0.1=py37_0
expat=2.2.6=he6710b0_0
fastcache=1.0.2=py37h14c3975_2
ffmpeg=4.0=hcdf2ecd_0
filelock=3.0.10=py37_0
flask=1.0.2=py37_1
flask-cors=3.0.7=py37_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
future=0.17.1=py37_0
geographiclib=1.50=pypi_0
geopy=2.0.0=pypi_0
get_terminal_size=1.0.0=haa9412d_0
gevent=1.3.7=py37h7b6447c_1
glib=2.56.2=hd408876_0
glob2=0.6=py37_1
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py37h10f8cd9_2
graphite2=1.3.13=h23475e2_0
greenlet=0.4.15=py37h7b6447c_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
heapdict=1.0.0=py37_2
html5lib=1.0.1=py37_0
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imageio=2.4.1=py37_0
imagesize=1.1.0=py37_0
importlib_metadata=0.6=py37_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py37h39e3cac_0
ipython=7.2.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.4.2=py37_0
isort=4.3.4=py37_0
itsdangerous=1.1.0=py37_0
jasper=2.0.14=h07fcdf6_1
jbig=2.1=hdba287a_0
jdcal=1.4=py37_0
jedi=0.13.2=py37_0
jeepney=0.4=py37_0
jinja2=2.10=py37_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.2.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.4.0=py37_0
jupyterlab=0.35.3=py37_0
jupyterlab_server=0.2.0=py37_0
keyring=17.0.0=py37_0
kiwisolver=1.0.1=py37hf484d3e_0
krb5=1.16.1=h173b8e3_7
lazy-object-proxy=1.3.1=py37h14c3975_2
libarchive=3.3.3=h5d8350f_5
libcurl=7.63.0=h20c2e04_1000
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
liblief=0.9.0=h7725739_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.0=h1ba5d50_4
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libtool=2.4.6=h7b6447c_5
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
libxslt=1.1.32=h1312cb7_0
llvmlite=0.26.0=py37hd408876_0
locket=0.2.0=py37_1
lxml=4.2.5=py37hefd8a0e_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
markupsafe=1.1.0=py37h7b6447c_0
matplotlib=3.0.2=py37h5429711_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2019.1=144
mkl-service=1.1.2=py37he904b0f_5
mkl_fft=1.0.10=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
more-itertools=4.3.0=py37_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py37_0
msgpack-python=0.5.6=py37h6bb024c_1
multipledispatch=0.6.0=py37_0
navigator-updater=0.2.1=py37_0
nbconvert=5.4.0=py37_1
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
networkx=2.2=py37_1
ninja=1.10.2=hff7bd54_1
nltk=3.4=py37_1
nose=1.3.7=py37_2
notebook=5.7.4=py37_0
numba=0.41.0=py37h962f231_0
numexpr=2.6.8=py37h9e4a6bb_0
numpy=1.16.2=py37h7e9f1db_0
numpy-base=1.16.2=py37hde5b4d6_0
numpydoc=0.8.0=py37_0
odo=0.5.1=py37_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openpyxl=2.5.12=py37_0
openssl=1.1.1k=h27cfd23_0
packaging=18.0=py37_0
pandas=0.23.4=py37h04863e7_0
pandoc=1.19.2.1=hea2e7c5_1
pandocfilters=1.4.2=py37_1
pango=1.42.4=h049681c_0
parso=0.3.1=py37_0
partd=0.3.9=py37_0
patchelf=0.9=he6710b0_3
path.py=11.5.0=py37_0
pathlib2=2.3.3=py37_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
pep8=1.7.1=py37_0
pexpect=4.6.0=py37_0
pickleshare=0.7.5=py37_0
pillow=5.3.0=py37h34e0f95_0
pip=19.2.3=py37_0
pixman=0.38.0=h7b6447c_0
pkginfo=1.4.2=py37_1
pluggy=0.8.0=py37_0
ply=3.11=py37_0
prometheus_client=0.5.0=py37_0
prompt_toolkit=2.0.7=py37_0
psutil=5.4.8=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
pulp=2.3.1=pypi_0
py=1.7.0=py37_0
py-lief=0.9.0=py37h7725739_1
py-opencv=3.4.2=py37hb342d67_1
pycodestyle=2.4.0=py37_0
pycosat=0.6.3=py37h14c3975_0
pycparser=2.19=py37_0
pycrypto=2.6.1=py37h14c3975_9
pycurl=7.43.0.2=py37h1ba5d50_0
pyflakes=2.0.0=py37_0
pygments=2.3.1=py37_0
pylint=2.2.2=py37_0
pyodbc=4.0.25=py37he6710b0_0
pyopenssl=18.0.0=py37_0
pyparsing=2.3.0=py37_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.6.8=py37_0
pytables=3.4.4=py37ha205bf6_0
pytest=4.0.2=py37_0
pytest-arraydiff=0.3=py37h39e3cac_0
pytest-astropy=0.5.0=py37_0
pytest-doctestplus=0.2.0=py37_0
pytest-openfiles=0.3.1=py37_0
pytest-remotedata=0.3.1=py37_0
python=3.7.4=h265db76_1
python-dateutil=2.7.5=py37_0
python-libarchive-c=2.8=py37_6
pytorch=1.9.0=py3.7_cuda11.1_cudnn8.0.5_0
pytz=2018.7=py37_0
pywavelets=1.0.1=py37hdd07704_0
pyyaml=3.13=py37h14c3975_0
pyzmq=17.1.2=py37h14c3975_0
qt=5.9.7=h5867ecd_1
qtawesome=0.5.3=py37_0
qtconsole=4.4.3=py37_0
qtpy=1.5.2=py37_0
readline=7.0=h7b6447c_5
requests=2.21.0=py37_0
rope=0.11.0=py37_0
ruamel_yaml=0.15.46=py37h14c3975_0
scikit-image=0.14.1=py37he6710b0_0
scikit-learn=0.20.1=py37hd81dba3_0
scipy=1.1.0=py37h7c811a0_2
seaborn=0.9.0=py37_0
secretstorage=3.1.0=py37_0
send2trash=1.5.0=py37_0
setuptools=41.2.0=py37_0
simplegeneric=0.8.1=py37_2
singledispatch=3.4.0.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
snappy=1.1.7=hbae5bb6_3
snowballstemmer=1.2.1=py37_0
sortedcollections=1.0.1=py37_0
sortedcontainers=2.1.0=py37_0
sphinx=1.8.2=py37_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-websupport=1.1.0=py37_1
spyder=3.3.2=py37_0
spyder-kernels=0.3.0=py37_0
sqlalchemy=1.2.15=py37h7b6447c_0
sqlite=3.29.0=h7b6447c_0
statsmodels=0.9.0=py37h035aef0_0
sympy=1.3=py37_0
tblib=1.3.2=py37_0
terminado=0.8.1=py37_1
testpath=0.4.2=py37_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py37_0
torchaudio=0.9.0=py37
torchvision=0.2.2=py_3
tornado=5.1.1=py37h7b6447c_0
tqdm=4.28.1=py37h28b3542_0
traitlets=4.3.2=py37_0
typing_extensions=3.7.4.3=pyha847dfd_0
unicodecsv=0.14.1=py37_0
unixodbc=2.3.7=h14c3975_0
urllib3=1.24.1=py37_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.14.1=py37_0
wheel=0.33.6=py37_0
widgetsnbextension=3.4.2=py37_0
wrapt=1.10.11=py37h14c3975_2
wurlitzer=1.0.2=py37_0
xlrd=1.2.0=py37_0
xlsxwriter=1.1.2=py37_0
xlwt=1.3.0=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.2.5=hf484d3e_1
zict=0.1.3=py37_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
